Moving Controllers Around Moving results page controller to UserAccountController file, and renaming results.html to userdash.html, since the plan will be to use this page for user dashboard functionality (#8 and #24)
